+++ android-party/feature-login/src/main/res/drawable/rectangle_button.xml
@@ -1,11 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <solid android:color="@color/button_green"/>
-
-    <corners
-        android:bottomLeftRadius="7dp"
-        android:bottomRightRadius="7dp"
-        android:topLeftRadius="7dp"
-        android:topRightRadius="7dp"/>
-</shape>
+<ripple xmlns:android="http://schemas.android.com/apk/res/android"
+    android:color="@color/button_green_dark">
+    <item>
+        <shape android:shape="rectangle">
+            <solid android:color="@color/button_green" />
+            <corners android:radius="7dp" />
+        </shape>
+    </item>
+</ripple>
