+++ android-party/feature-login/src/main/res/layout-land/activity_login.xml
@@ -99,6 +99,5 @@
             app:layout_constraintEnd_toEndOf="@+id/password"
             app:layout_constraintStart_toStartOf="@+id/password"
             app:layout_constraintVertical_bias="0.3" />
-
     </androidx.constraintlayout.widget.ConstraintLayout>
 </FrameLayout>
