+++ android-party/feature-login/src/main/res/values/colors.xml
@@ -8,6 +8,7 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="button_green">#9fd533</color>
+    <color name="button_green_dark">#8EBF2B</color>
     <color name="light_black">#191917</color>
     <color name="gray_200">#F0F0F0</color>
 </resources>
