+++ android-party/feature-server/src/main/java/com/example/feature_server/presentation/view/activity/ServerActivity.kt
@@ -3,6 +3,9 @@ package com.example.feature_server.presentation.view.activity
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupWithNavController
@@ -11,6 +14,7 @@ import com.example.feature_server.databinding.ActivityServerBinding
 import com.example.feature_server.presentation.viewmodel.ServerContract
 import com.example.feature_server.presentation.viewmodel.ServerViewModel
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class ServerActivity : AppCompatActivity() {
@@ -26,9 +30,32 @@ class ServerActivity : AppCompatActivity() {
         val navController = navHostFragment.navController
         val appBarConfiguration = AppBarConfiguration(navController.graph)
         binding.toolbar.setupWithNavController(navController, appBarConfiguration)
+        observeEffect()
+    }
+
+    private fun observeEffect() {
+        lifecycleScope.launch {
+            repeatOnLifecycle(Lifecycle.State.STARTED) {
+                viewModel.effect.collect { effect ->
+                    when (effect) {
+                        is ServerContract.Effect.OnLogoutEffect -> {
+                            logout()
+                        }
+                        else -> {
+                            // do nothing
+                        }
+                    }
+                }
+            }
+        }
     }
 
     override fun onBackPressed() {
         viewModel.onUiEvent(ServerContract.Event.OnLogoutClicked)
+    }
+
+    private fun logout() {
+        finish()
+        overridePendingTransition(android.R.anim.slide_in_left, android.R.anim.slide_out_right)
     }
 }
