+++ android-party/feature-server/src/main/res/layout/header_layout.xml
@@ -3,7 +3,9 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content">
+    android:layout_height="wrap_content"
+    android:elevation="8dp"
+    android:background="@drawable/header_rect">
 
     <TextView
         android:id="@+id/name"
@@ -11,7 +13,7 @@
         android:layout_height="60dp"
         android:padding="16dp"
         android:text="@string/server"
-        android:textSize="@dimen/header_text_size"
+        android:textSize="@dimen/text_size"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
@@ -25,7 +27,7 @@
         android:gravity="start"
         android:padding="16dp"
         android:text="@string/distance"
-        android:textSize="@dimen/header_text_size"
+        android:textSize="@dimen/text_size"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="@+id/guideline2"
