+++ android-party/feature-server/src/main/res/navigation/nav_graph.xml
@@ -10,7 +10,11 @@
         android:label="LoadFragment" >
         <action
             android:id="@+id/action_loadFragment_to_serverFragment"
-            app:destination="@id/serverFragment" />
+            app:destination="@id/serverFragment"
+            app:enterAnim="@anim/slide_in_right"
+            app:exitAnim="@anim/slide_out_left"
+            app:popEnterAnim="@anim/slide_in_right"
+            app:popExitAnim="@anim/slide_out_left" />
     </fragment>
     <fragment
         android:id="@+id/serverFragment"
